convert accesslist + use paginated endpoint
